Phonograph: fix MedisStore access for artwork
its all LOOOOONG values - this broke artwork access
also move from getExternalStorageDir to getExternalFilesDir
and make it buildable again in android studio

+++ app/src/main/java/com/kabouzeid/gramophone/model/Album.java
@@ -20,7 +20,7 @@ public class Album implements Parcelable {
         this.songs = new ArrayList<>();
     }
 
-    public int getId() {
+    public long getId() {
         return safeGetFirstSong().albumId;
     }
 
@@ -28,7 +28,7 @@ public class Album implements Parcelable {
         return safeGetFirstSong().albumName;
     }
 
-    public int getArtistId() {
+    public long getArtistId() {
         return safeGetFirstSong().artistId;
     }
